change data tyoe if total in ticket from float to double

+++ src/main/java/com/fzoo/zoomanagementsystem/model/Ticket.java
@@ -23,6 +23,6 @@ public class Ticket {
     private int quantity;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate date;
-    private float total;
+    private double total;
     private int account_id;
 }
